Fix(unittest_json): Change order of JSON keys
The JSON keys are now ordered as follows:

```
{
  "status": "fail",
  "message": null,
  "tests": [
    {
      "name": "Test that the thing works",
      "status": "fail",
      "message": "Expected 42 but got 123123",
      "output": "Debugging information output by the user"
    }
  ]
}
```

which is the JSON structure from:
- https://github.com/exercism/automated-tests/blob/master/docs/interface.md

We currently omit a `message` key (rather than setting its value to
`null`) if its sibling `status` key has a value that is not `error`.

However, the per-test `output` key is currently always included and
always set to the empty string.