Fix(runner): Fix importing of `unittest_json`
We made a processed exercise test file able to import `unittest_json`
simply by copying `unittest_json.nim` to the temporary directory.
However, we did this copying only when running the existing tests, and
not when executing the `runner` binary.

Therefore this commit moves the copying of `unittest_json.nim` into
`runner.nim`.

We will add tests for the `runner` binary soon.