Test: Add `pass/multiple_tests/` case
This commit adds our second test case, which is a template for future
ones.

For readability, we want the tested function to:
1. Be simple. This will help distinguish the tests that are designed to
   fail. It can also simplify failure messages.
2. Be consistent for all test cases. The existing `pass/single_test/`
   case is `hello-world`, but that can stay for now - we test that
   `prepareFiles` prepares it as expected.
3. Have at least one input parameter. This is useful for later test
   cases.
4. Return an `int`. Later test cases have at least 3 `test` lines,
   and other basic types like `bool`, `enum`, `char` and `string`
   probably produce less readable test files and failure messages.

So let's use the identity function - it's a simple function that
satisfies the above.

Some candidates rejected for being arguably less readable:
- `func plus*(a, b: int): int = a + b`
- `func plusOne*(n: int): int = n + 1`
- `func double*(n: int): int = n * 2`
- `func square*(n: int): int = n * n`

The first is rejected because we probably don't need to add a test case
later that uses more than one input parameter. And for maximal
readability, we'd set the first parameter to 0 (in which case we might
as well use `identity`).

Other decisions:
- Use `func` - we can switch to `proc` for later test cases that have
  side effects.
- Don't add a `# version 1.0.0` or similar to the test file. It would
  have no meaning, and could be confusing.
- Name the function `identity`, not e.g. `i`, `id`, or `I`.
- Name the module `identity`, not e.g. `i`, `id` or `identity_function`.