Test: Add `error/compiletime_crash_in_solution/` case
This commit adds a test for the case of the student's solution code
causing an internal compiler error.

The crash-triggering code is chosen because:
- It's short.
- It's easy to understand.
- The relevant GitHub issue has been open for a while (since July 2019).

We can replace the code if the issue gets fixed.

We will need to update the `message` contents in `expected_results.json`
after Nim 1.4.0 is released. This is because the `devel` Nim 1.3.5
produces a slightly different message than the stable Nim 1.2.2
currently used by our CI and Docker image.