Test: Add `pass/single_test_output_stderr_only`
The previous commit changed the unicode output case to no longer be
`stderr`-only, so we might as well add a stderr-only test case.

It might be better to make all the other output test cases be
stdout-only, and perhaps have one that tests both, but we can consider
that later.