Build: Make it easier to bump the Nim version (#40)
This commit refactors the CI workflow and the Dockerfile so that each
defines the desired Nim version only once.

For `ci.yml` it does this by changing the testing-only job to install
Nim directly, rather than using a container with the `ubuntu-regular`
image from `nimlang/nim`. This also makes it more obvious that we're
using the `alpine-slim` image as the base for the image we build.