CI: Add support for Exercism v3 track structure
As part of the preparation for Exercism v3, the v2 exercises on all
tracks were moved a new subdirectory: `exercises/practice`.

In order to test our test runner, we run it on exercises from the Nim
track repo during CI. This commit updates our CI and tests to use the
new location for practice exercises, making CI green again.

See:
- https://github.com/exercism/nim/commit/f7ab39d96e32