ci: add linting of Dockerfile
It is a good idea to lint our Dockerfile. `hadolint` is one such
linter (the name is short for "Haskell Dockerfile Linter") - it
also runs `shellcheck` internally.

Before this commit, running `hadolint` on our Dockerfile would produce:
```
Dockerfile:2 DL3006 warning: Always tag the version of an image explicitly
Dockerfile:7 DL3006 warning: Always tag the version of an image explicitly
Dockerfile:8 DL3018 warning: Pin versions in apk add. Instead of `apk
add <package>` use `apk add <package>=<version>`
```
The first two were false positives, and the third is a real issue.

Therefore this commit also alters the Dockerfile to workaround the false
positives, and to suppress the warning that is difficult to resolve.

See:
- https://github.com/brpaz/hadolint-action
- https://github.com/hadolint/hadolint