build: use `alpine` directly, not `nimlang/nim`
This commit reduces the size of our Docker image. We were using a
`nimlang/nim` image that is itself based off `alpine:latest`, but we can
just use an `alpine` image directly to avoid some bloat.

Before: 249.53 MB
After:  139.60 MB

Other advantages:
- Reduced attack surface: we get more control over the image contents -
  we can remove things that we don't need in the test runner
  environment, like `nodejs`.
- We are no longer dependent on `nimlang/nim` for Nim version updates.
- We can update Alpine even if there hasn't been a new Nim version.

Disadvantages:
- `docker build` now takes something like 5 minutes during CI, rather
  than something like 30 seconds.
- Some added complexity.

This commit removes:
- `nodejs`
- The C++ compiler. `nim cpp` is now impossible
- The ability to run `nim doc`
- The ability to run `nim js`
- Some build dependencies: `curl`, `tar`, `xz`
- Unnecessary Nim directories

With this commit, we now build Nim ourselves. We do this by adding a
script, which is better than inlining the commands in the Dockerfile
because we can:
- run and test the script independently.
- run `shellcheck` on it.

This commit also adds a `shellcheck` workflow.