build: use `alpine` for parent image, not `nimlang/nim` (#64)
This commit reduces the size of our Docker image. Previously, we were
using an image from the `nimlang/nim` Docker Hub repo that itself uses
`alpine:latest`. We now use an `alpine` image directly, thereby removing
some features/dependencies we don't need in the test runner environment.

Image size before this commit: 249.53 MB
Image size with this commit:   139.66 MB

Some advantages:
- Reduced attack surface
- Reduced time to download and extract the parent image
- Reduced time to transfer our built image from/to the container
  registry
- We can update Nim even if https://hub.docker.com/r/nimlang/nim wasn't
  updated yet
- We can update Alpine even if there hasn't been a new Nim version

Disadvantages:
- As we now build the Nim compiler ourselves, `docker build` now takes
  something like 5 minutes during CI, rather than something like 30
  seconds. We could add caching to the CI workflow if this becomes too
  inconvenient.
- Some added complexity.

This commit removes from the image:
- `nodejs`, and the ability to run `nim js` (compile Nim to JavaScript)
- `g++`, which removes the ability to run `nim cpp` (compile Nim to C++)
- The ability to run `nim doc` (generate documentation for a Nim file)
- Some build dependencies, for example: `curl`, `tar`, `xz`
- Unnecessary contents of the Nim installation

To build the Nim compiler, this commit adds a new script - this is
better than writing the commands in the Dockerfile directly because we
can:
- run and test the script independently
- run `shellcheck` on it

This commit also adds a `shellcheck` workflow.