.github(pause-community-contributions): add custom token (#146)
The pause-community-contributions workflow makes a call to the 
GitHub API to check if the person contributing is a member of the
organization. However, this call currently fails if the contributor has
set their membership to 'private'.

This is because the default token provided for GitHub Actions only has
permissions for the repository, not for the organization. With this
token, we're not allowed to see private memberships.

We've created a custom, org-wide secret containing a personal token that
has permissions to read organization membership. Unfortunately the
secret cannot be accessed directly by the shared workflow, it has to be
passed in.

We updated the shared workflow to use the token, if it is provided, and
this PR updates the workflow in this repo to pass the secret.

Also remove the workflow_dispatch, which would always fail on use.